- Added option to add custom types to Red4Reader - Some cleanup
