+++ WolvenKit.RED4.Save/Helper/ClassHashHelper.cs
@@ -5,100 +5,64 @@ namespace WolvenKit.RED4.Save;
 
 public class ClassHashHelper
 {
-    private static readonly Dictionary<ulong, Type> _classHashes = new();
-    private static Dictionary<Type, ulong> _classHashesReverse = new();
-
-    private static readonly Dictionary<ulong, RedReflection.ExtendedPropertyInfo> _propertyHashes = new();
-    private static readonly Dictionary<Type, ulong> _typeHashes = new();
+    private static readonly Dictionary<ulong, Type> s_classHashes = new();
+    private static Dictionary<Type, ulong> s_classHashesReverse = new();
 
 
     private static void GenerateClassHashes()
     {
         foreach (var redType in RedReflection.GetTypes())
         {
-            _classHashes.Add(FNV1A64HashAlgorithm.HashString(redType.Key), redType.Value);
+            s_classHashes.Add(FNV1A64HashAlgorithm.HashString(redType.Key), redType.Value);
         }
 
-        _classHashesReverse = _classHashes.ToDictionary(x => x.Value, x => x.Key);
+        s_classHashesReverse = s_classHashes.ToDictionary(x => x.Value, x => x.Key);
     }
 
     public static Type? GetTypeFromHash(ulong hash)
     {
-        if (_classHashes.Count == 0)
+        if (s_classHashes.Count == 0)
         {
             GenerateClassHashes();
         }
 
-        return _classHashes.TryGetValue(hash, out var result) ? result : null;
+        return s_classHashes.TryGetValue(hash, out var result) ? result : null;
     }
 
     public static ulong GetHashFromType(Type type)
     {
-        if (_classHashesReverse.Count == 0)
+        if (s_classHashesReverse.Count == 0)
         {
             GenerateClassHashes();
         }
 
-        return _classHashesReverse[type];
+        return s_classHashesReverse[type];
     }
 
-    public static RedReflection.ExtendedPropertyInfo? GetPropertyInfo(Type clsType, ulong propHash)
+    public static ExtendedPropertyInfo? GetPropertyInfo(Type clsType, ulong propHash)
     {
-        //if (_propertyHashes.ContainsKey(propHash))
-        //{
-        //    return _propertyHashes[propHash];
-        //}
-
-        var result = InternalGetProperty(clsType, propHash);
-        if (result == null)
+        var typeInfo = RedReflection.GetTypeInfo(clsType);
+        foreach (var propertyInfo in typeInfo.PropertyInfos)
         {
-            return null;
-        }
-
-        return result;
-
-
-        RedReflection.ExtendedPropertyInfo? InternalGetProperty(Type clsType, ulong propertyHash)
-        {
-            var typeInfo = RedReflection.GetTypeInfo(clsType);
-            foreach (var propertyInfo in typeInfo.PropertyInfos)
+            if (string.IsNullOrEmpty(propertyInfo.RedName))
             {
-                if (string.IsNullOrEmpty(propertyInfo.RedName))
-                {
-                    continue;
-                }
-
-                var hash = FNV1A64HashAlgorithm.HashString(propertyInfo.RedName);
-                //if (!_propertyHashes.ContainsKey(hash))
-                //{
-                //    _propertyHashes.Add(hash, propertyInfo);
-                //}
-                
-                if (propHash == hash)
-                {
-                    return propertyInfo;
-                }
+                continue;
             }
 
-            if (typeInfo.BaseType != typeof(RedBaseClass))
+            if (FNV1A64HashAlgorithm.HashString(propertyInfo.RedName) == propHash)
             {
-                return InternalGetProperty(typeInfo.BaseType, propertyHash);
+                return propertyInfo;
             }
-
-            return null;
         }
+
+        if (typeInfo.BaseType != typeof(RedBaseClass))
+        {
+            return GetPropertyInfo(typeInfo.BaseType, propHash);
+        }
+
+        return null;
     }
 
-    public static ulong GetRedTypeHash(RedReflection.ExtendedPropertyInfo propertyInfo)
-    {
-        //if (_typeHashes.ContainsKey(type))
-        //{
-        //    return _typeHashes[type];
-        //}
-
-        var hash = FNV1A64HashAlgorithm.HashString(RedReflection.GetRedTypeFromCSType(propertyInfo.Type, propertyInfo.Flags));
-        //_typeHashes.Add(type, hash);
-
-        return hash;
-    }
+    public static ulong GetRedTypeHash(ExtendedPropertyInfo propertyInfo) =>
+        FNV1A64HashAlgorithm.HashString(RedReflection.GetRedTypeFromCSType(propertyInfo.Type, propertyInfo.Flags));
 }
